requirements.txt
fastapi
pillow
qwen_vl_utils
torch
torchvision
transformers
uvicorn